+++ apps/ymir-sdl3/src/app/ui/views/settings/video_settings_view.cpp
@@ -241,14 +241,32 @@ void VideoSettingsView::Display() {
 
     widgets::settings::video::enhancements::Deinterlace(m_context);
     widgets::settings::video::enhancements::TransparentMeshes(m_context);
+    widgets::settings::video::enhancements::ResolutionScaling(m_context);
 
     // -----------------------------------------------------------------------------------------------------------------
 
     ImGui::PushFont(m_context.fonts.sansSerif.bold, m_context.fontSizes.large);
+    ImGui::SeparatorText("Rendering");
+    ImGui::PopFont();
+
+    widgets::settings::video::UseHardwareAcceleration(m_context);
+
+    // -----------------------------------------------------------------------------------------------------------------
+
+    ImGui::PushFont(m_context.fonts.sansSerif.bold, m_context.fontSizes.medium);
     ImGui::SeparatorText("Software renderer");
     ImGui::PopFont();
 
     widgets::settings::video::swrenderer::ThreadedVDP(m_context);
+
+    // -----------------------------------------------------------------------------------------------------------------
+
+    ImGui::PushFont(m_context.fonts.sansSerif.bold, m_context.fontSizes.medium);
+    ImGui::SeparatorText("Hardware renderer");
+    ImGui::PopFont();
+
+    widgets::settings::video::hwrenderer::VDP1VRAMSyncMode(m_context);
+    widgets::settings::video::hwrenderer::VDP2VRAMSyncMode(m_context);
 }
 
 } // namespace app::ui